renamed rest -> service as files have changed location
